put the run output processing into a module
