+++ bin/VERSION
@@ -1,1 +1,1 @@
-0.3.2 (built 2026-07-09)
+0.3.2 (built 2026-07-10)
